+++ CHANGELOG.md
@@ -10,9 +10,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 ### Changed
-* Changed and update the `compas_view2` examples into `compas_viewer`.
+
+* Changed and updated the `compas_view2` examples into `compas_viewer`.
+* Changed `compas.scene.Scene` to inherent from `compas.datastructrues.Tree`.
+* Changed `compas.scene.SceneObject` to inherent from `compas.datastructrues.TreeNode`.
 
 ### Removed
+
+* Removed `compas.scene.SceneObjectNode`, functionalities merged into `compas.scene.SceneObject`.
+* Removed `compas.scene.SceneTree`, functionalities merged into `compas.scene.Scene`.
 
 
 ## [2.1.0] 2024-03-01